Change how ClrRuntime is created
This fixes a longstanding design flaw in ClrMD:  we construct a
ClrRuntime based only on a dac file.  The correct choice is to let the
user select exactly what version of CLR they want to debug, and have the
dac location be an optional parameter.  (If no dac location is
specified, we will search for it on the local machine and symbol
servers.)

The new code pattern looks something like this:
ClrRuntime runtime = dataTarget.ClrVersions[?].CreateRuntime();