Background fetch binaries that will be needed
In minidumps, we will need to hit the symbol server often to get managed
binaries.  This is because metadata (needed for a lot of operations) is
baked into the image that may not be in the crash dump.  This change
makes it so that we start fetching those binaries on a background thread
as soon as we find that we are missing metadata in any managed module.
This should speed up response time for a lot of operations.