Be more resilient to AppDomain enumeration issues
If we fail to enumerate the assemblies in an AppDomain, we should still
try to construct as much as we can.  This happens in cases where the
runtime is in an inconsistent state (no real bug, just the process was
stopped when the list of assemblies in an AppDomain is not walkable).