Further error hardening for assemblies and modules
We were returning null in failure cases of v4.5's GetAssemblyList and
GetModuleList.  This change removes the return value check from these
methods, as the list it generates can be partially filled when the dac
hits an error.  The caller of these methods handles null values
gracefully, so this enables us to construct AppDomain and module lists
in places we couldn't before.