Define a symbol provider for .Net Native
ClrMD uses PDBs for .Net Native.  This checkin is the initial step for
allowing the user to provide symbols support.

The primary concern with having ClrMD read the PDBs itself is that other
parts of the tool may use PDBs as well...leading us to load them twice
in memory.  This gives the developer more control over memory in the
process.