Rename EnumerateObjects to EnumerateObjectAddresses
This change renames EnumerateObjects to EnumerateObjectAddresses.  ClrMD
will support enumerating a strongly typed "ClrObject" in the future (in
addition to the lighter weight EnumerateObjectAddresses), and the name
will conflict with what will be eventually implemented.

This should be the last of the API surface area changes for v1.0 unless
I missed something.