Set version to match NuGet
Set the version to match 0.8.31.0, which is the NuGet versioning system.

I used to carry two different versions, one for the NuGet package (which
I have little control over) and one for the real version I keep track
of.  Now we will only track the NuGet release version (though it is
lower than the one we were using...).