Get ICorDebugILFrame for ClrStackFrames
This change adds DesktopStackFrame.CordbFrame, which is its
ICorDebugILFrame.  This is now properly filled.  This will allow us to
get local variables on demand from ICorDebug.
I also cleaned up the use of _CONTEXT structures in ClrMD to use a
thread static byte array and not maintain magic constants in the actual
code.  (ContextHelper still contains the magic constants though.)