Fix issue looking up metadata in triage dumps
In triage dumps, we don't have the memory in the crash dump to properly
get the file name of a PEFile loaded into memory.  As a result, we tear
down the process trying to prefetch metadata for the binary due to an
exception on a background thread.  This fixes the immediate problem by
ensuring PrefetchBinary never tears down the process (and null-checking
the file name).

In the future we need to try to look up the binary name in the crash
dump instead, but that is not implemented in this change.