Move to new ClrFlavor for .Net Core
'coreclr' as a dll is actually a set of forks over a similar codebase.
Some of these have incompatibilities which will need to be handled
separately.  ClrMD will be supporting .Net Core as a debuggee (not yet
running as a .Net Core library itself), but some care has to be taken to
call out that we are not supporting Silverlight or the version of
'coreclr' that ships with .Net Native at this time.