Mark .Net Native obsolete in this library
I plan to strip .Net Native support from ClrMD and move that into a smaller, more targeted library.  Originally I intended ClrMD to be a catch-all api to debug any CLR process, but unfortunately .Net Native is sufficiently different that this was a bad idea from the start.  In fact, only a handful of things actually work in the .Net Native support (such as heap walking, which was built for PerfView).

This change only marks .Net Native as obsolete as an FYI that this change is coming (and this checking notice for others to comment and weigh in before I actually make the change).  In the future, ClrMD will simply stop enumerating .Net Native runtimes in DataTarget.ClrVersions.