Add new BaseDesktopHeapType class named PrimitiveType to host ValueTypes such as Int32/Int16/Bool/Char/Pointer/Etc.
Modify GetBasicType() to return a PrimitiveType object when InitBasicTypes() fails to load the type from mscorlib.  This happens in all MiniDumps.

All the information to be able to extract this data is available to ClrMD, this patch allows us to get this Basic information even when InitBasictypes() can not find the types in a MiniDump.

This allows us to get the same information !sos.do gets in Minidumps.