Implement .Net Core build and x64 Linux support (#123)
This change adds .Net Core build support (next NuGet release will include .Net Standard 2.0 library). This change also adds support for Linux core dump debugging. Attaching to live processes on Linux not implemented in this build.

I'm pushing this down to the main branch to stabilize, merge other PRs, and get it ready to ship. Please do not consider this build final and the surface area stable until this is shipped as a NuGet package.  There are likely some issues right now, but I'm working to resolve them asap.

Some high level details:

Moved the build to Arcade to support building on linux.
Disabled tests, as they do not function under Arcade. .Net Core does not support CodeDOM, which the current test system uses as well. I'll have to rethink/reimplement the test infrastructure in a future change. The tests do pass when I forced them to run in an outside environment.
Removed COM wrappings of dac interfaces and reimplement them by hand. (.Net Core on linux doesn't have COM enabled, and CLR still uses it on Linux...)
Added an Elf core dump reader. This is not intended to ever be public.
Shuffled around some of the guts of ClrMD, but no functionality should be changed or affected.