Various cleanups (#133)
* Rename ComWrappers to "DacInterface"

* Change some access for classes to support internal tooling

* Remove accidental addition of license file

* Ensure DacLibrary is not freed when there is any COM interface in use

* Fix path issue introduced by a previous PR

* Cleanup tests, move to xunit

This is not currently functional, only enabling clean compile.

* Fix some over/underflow exceptions

* Don't run the release code if we don't succeed the QueryInterface call

* Fix string marshalling issue

* Fix some incorrect wrappings

* Stop asserting on s_totalInstanceCount > 0

This is a bit concerning.  We apparently are hitting an s_totalInstanceCount <= 0.  This shouldn't happen, but I suspect the count is incorrect here.

* Fix NullRef when we can't get method props

If IMetaDataImport is null don't throw a null-ref.

* Remove some Console.WriteLines

* Fix issue with handle enumeration

* Make tests work...

Updated code so that the tests for ClrMD are mostly functional.  There's still an outstanding test failure with FQ tests, but I'm not sure if that's a faulty test or a problem with ClrMD yet.

* Fix FinalizationQueue test

The finalizer test wasn't structured in a way to properly collect references when built in the new system.

* Change HeapDetails marshalling to use arrays

* Remove Console.WriteLine

* Startup path cleanups

* Make module address public

* Add GetMethodDescFromToken

* Rename a few variables to be more consistent

* Remove duplicated build.cmd with different casing