Final changes for 1.0 (#146)
Now that we are marking the library 1.0, we needed to remove some Obsolete code markers. We also completely stripped support for .Net Native. The intent is to have a new Nuget package (maybe Microsoft.Diagnostics.Runtime.Native.dll) to use instead. This new library is not yet complete or published to Nuget.

* Split .Net Native into its own library

- Split .Net Native into its own library.
- The implementation is incomplete, but I needed to split it now before releasing 1.0-prerelease.

* Remove some obsolete code

* Remove .Net Native ClrFlavor