Set sensible limits to stack root enumeration
If we get a bad TEB or otherwise mis-calculate a stack's base/limit, we
can end up walking forever.  This checkin puts a limit (1gb) on how much
we will walk on a stack in case we somehow got a bad answer for the
range of memory we calculated.

This is a workaround for https://github.com/Microsoft/clrmd/issues/89.