Fix LinuxFunctions.GetProcAddress throwing an exception (#274)
Because it uses System.Runtime.InteropServices.NativeLibrary.GetExport instead of TryGetExport
it throws an exception when the function can't be found instead of return null. This breaks
the DacLibrary init code (DacLibrary.cs line 86) when it is try to get the DAC_PAL_InitializeDLL
export and then the PAL_InitializeDLL export.