Various fixes
Change thread CONTEXT size check to return error if less than instead of not equal.

Make COM interop wrapper accessible

Add setter to ModuleInfo.IsManaged and change the ELF dump reader
to set it if not an ELF module assuming it is a PE file.

Add auxv structs/enums. Add module list filtering for the "interpreter"
and any module name that begins with "\dev\".