Allow ClrMD to be used from multiple threads (#376)
Initial implementation of support for parallel processing. Not all data targets have been marked as threadsafe yet. Only DbgEngDataReader and Windows's LiveDataReader are currently thread safe.

Note that we do not expect to see any perf improvements by running ClrMD operations in parallel. This is because the underlying diagnostics API mscordac*.dll takes a giant lock around any call we make, and many IDataTarget implementations will similarly need to take a lock around from the target.

Using ClrMD from multiple threads is primarily for convenience. That is, not having to keep ClrMD locked to a single thread, you can now use it with in an async/await context, for example. This also allows parallel processing for user code if you are in a scenario where you only need to read from the dac/data target a little and do a lot of processing yourself.

Changes:

* Ensure we only ever produce a single heap
* Allow recreating ClrRuntime
* Make MemoryReader a thread static
* Add parallel stress test
* Add .xml to the ignore list to stop checking the files in
* Clean up ParallelStressTest
* Inline size calculation
* Add heap walk logging
* Rename IsMinidump
* Add IDataReader.IsThreadSafe
* Add ClrRuntime/ITypeFactory.IsThreadSafe
* Lock cache before clearing it