Fix sign extension issues (#419)
* Fix sign extension issues

CLRDATA_ADDRESS is defined as a signed 64 bit integer in the CLR codebase.  When inspecting a 32bit process, this means that we cannot simply use ulong to represent these values when communicating with the dac because addresses with the highest bit set will be sign extended by the dac.

This change adds a ClrDataAddress struct to represent a possibly sign-extended data value.  All structs that the dac produces have been updated to match their definition in coreclr/src/inc/dacprivate.h.

This struct will do the right thing to un-sign extend the address when inspecting a 32bit process.

* Fix more sign extension issues