Improve ELF reader EnumerateModules() perf
The dotnet-dump's `lm` command and any command run for the first time has a huge perf problem. The modules are enumerated to find the runtime module. Getting the Version for each PE image in the Linux core dump was taking way too long and failing.

The problem is the "layout" of managed PE modules. Some PE's are "file" layout and some are "loaded" (sections are aligned) layout in memory. The "isVirtual" flag in the PEImage constructor needs to be set based on the layout, but there is no way to determine that at the core dump data reader level. If the layout/isVirtual flag is set incorrectly, PEImage doesn't enumerate through the resources correctly, takes a long time because it is reading really long garbage strings, etc.

IDataReader.EnumerateModules() is meant to return just native modules (like libcoreclr.so, etc) but the PE's show up because it uses the NT_FILE entries in the coredump and because the PE's are memory mapped by the runtime as shared named memory,

The first part of the fix is to skip all the PE's (modules that ELFLoadedImage.Open can't open). The managed assemblies can be enumerated at a higher level (ClrRuntime.Modules) without any problems. dotnet-dump has a "clrmodules" command to do that.

The second part of the fix is to get the module's Version on demand and try both file and loaded layout of the PE images.

Fixed a problem in the ResourceEntry where resource id entries were trying to read a name string. Added check in the ResourceEntry read string code to limit strings to MaxPath (1024) characters for when the layout flag is wrong.

Fix PEImageResourceTests.FileInfoVersionTest. It wasn't comparing the same type of versions.

Issue: https://github.com/microsoft/clrmd/issues/564