Stop caching RCW/CCW results
We previously would cache whether a type was an RCW/CCW which in theory should work.  However, the dac private API only allows us to query for CCW/RCW results per object.  This means that sometimes when an object that was a CCW was already cleaned up, we'd mark the type permantently as not being a CCW when it was.  This lead to future calls to IsCCW/GetCCW to fail.