Temporarily disable test asset building
Building test assets with "AfterTargets=Build" is currently breaking the arcade build system signing step. This only happens within Arcade and no one on that team has been able to figure out why. Disabling this part of the build for now so we can get 2.0-beta builds out the door.