Don't filter out managed assemblies in Linux data reader.
It was a mistake to filter out the managed assemblies out of the ELF reader module list
because there are coredumps generated with the 2.1.x createdump that have a bug (which
will never get fixed because the service release bar for 2.1 is high) where the ELF module
can't be opened (ElfLoadedImage.Open() returns null). This causes most of the native modules
(especially libcoreclr.so) to be missing. With this fix most dotnet-dump/SOS commands will
work on this coredumps.

This is an regression caused by my module filtering change.