Don't throw an exception if we fail to QueryInterface IMetaDataImport (#686)
It's common for the QueryInterface for IMetaDataImport to fail if the dump file does not contain all of the metadata.  This is especially common in non-heap dumps and Watson dumps.  This change adds an extra QueryInterface to make sure this call will succeed before trying to create a MetadataImport instance.  This should greatly reduce the number of exceptions being thrown under normal operation.

Contributes to https://github.com/microsoft/clrmd/issues/416.