Fix IndexOutOfRangeException when getting fields
Occasionally we encounter a type where the reported MethodTableData.ParentMethodTable is non-null but invalid.  I'm not sure what causes that.  When we encounter these, it causes the field enumeration to fail because it will return the wrong fields for this type.  In these cases we'll just say there's no fields for that type since we cannot enumerate them correctly anyway.