Refactor cache entries (#752)
* Refactor cache entrys, specifically:

Make a common base class that represents most of the complexity that is
shared betweem the ArrayPool and AWE cache entries. Most importantly lock
management.

Simplify the locking scheme. Previously if you made a read that required
accessing N pages of data I would require locking all N pages of data and
ensuring they were all paged in before continuing with the read. I now
stream the reads so we lock/page-in one page of the read at a time. This
allows better granularity of concurrent reads as well as drastically simplifes
lock management. It also allows us to trim a cache entry we are currently
reading from as I don't announce a page-in completed (that could trigger a trim)
until AFTER I have read the data.

Fix a few bugs I found along the way around accurate cache entry size tracking
and the cache trimming logic

* Fix a couple instances of not being nullable aware (by disabling the check
in this specific code)

* Eliminate stupid null ref :P