Use custom PE image reader for coredumps
PEReader cannot handle only reading PEImages where the barest amount of the header is in a coredump.  It expects to be able to read deeply into the header to get get sections which might not be there.  The only place where this actually matters is the coredump reader, where we might not have the PE image laying around to read.

This change implements a mini PE header reader to get the filesize/timestamp without having to use PEReader.