Fixes linux dumps with module that is not a ELF or PE file in module list (#810)
Dumps generated on distro OpenSuse in the diagnostics repo testing have a weird module in
the image list: /usr/share/icu/52.1/icudt52l.dat. It is neither an ELF image or an PE file.

The call to PEImage.ReadIndexProperties throws an IndexOutOfRangeException exception in CoreDumpReader.CreateModuleInfo
because the buffer in PEImage.Read<T> is empty.