Fix ILOffsetMap in Minidumps
The dac reads a single byte of memory out of the instruction pointer you pass to EnumerateMethodInstancesByAddress to see if it's a valid location or not.  Unfortunately in a Minidump scenario it's normal for that read to fail since we simply did not put the addres into the dump file.

This change forces that single read to succeed.