Catch exception when parsing dac names
Wrap Dac name parsing in a try/catch.  Historically parsing all of the various type names has been a real pain point.  We need to fix the underlying exception but for now we'll wrap the method in a try/catch to ensure it doesn't unexpectedly throw.

Contributes to https://github.com/microsoft/clrmd/issues/897.