Fixes crash with ELF alpine arm32 dump
The unhandled exception in the ELF reader happened because there is a duplicate file pointer
entry for some images in the list. Removed the unnecessary ElfLoadedImage file table list
and tracked the minimum pointer instead. This is the symstore ELF reader does and it works
great.

Issue: https://github.com/dotnet/diagnostics/issues/3003

Passes the SOS tests.