Implement some DbgEng interfaces
This is a work in progress commit that I'm pushing to GitHub to make sure I don't lose my place.

Microsoft.Diagnostics.DbgEng is temporary, I'm probably going to push all of this code into the old Microsoft.Diagnostics.Runtime.Utilities namespace and start publishing that as a .Net 6+ nuget package.

The end goal here is to push all DbgEng based stuff out of ClrMD but have a more permanent (and supported) way to build DbgEng/ClrMD based plugins.

That's the plan so far anyway...it's not final.