Code review feedback
Change to do a first pass of only the first 8 1K chunks of each section and if that doesn't find the
dylinker header then go through every section 1K at time. This does significantly increases performance
because there are dumps that the dylinker header is in one of the very last section (i.e. a .NET 6.0 x64
Ventura MacOS dump with approx 4600 total sections).