Add a way to enumerate native heaps (#1072)
* Add a way to enumerate native heaps

Added ClrRuntime.EnumerateClrNativeHeaps.  This function will enumerate the native heaps/memory that CLR has allocated.  This is intended to be used in functions similar to SOS's !eeheap.

Added ISOSDacInterface13 to implement this and related functions.

Worked around an issue in .Net 7.

* Correct behavior on DesktopCLR

* Update PublicAPI.Shipped.txt

* Fix logic and invert else if/else

Inverted an else if/else to make it explicit that .Net Core 7 was the odd one out.