Change how we walk stackrefs/handles
There seems to be a bug in the dac when we pass a buffer less than the total amount of stack references or handles.

The original SOS C++ GCRoot code seems to have worked around it by finding the total number of entries and passing a buffer big enough to hold it.