Add cDac runtime enumeration support
Adds a new DebugLibaryKind.CDac enum used when the data contract export is found on a runtime module.

Currently the file path to the cdac is the same directory as the clrmd module.

Currently leaves it to the client (SOS in this case) to decide when it wants to instantiate the cdac. The standalone hosting support doesn't currently pick the cdac.

Breaking changes:

Add "bool verifySignature" to IClrInfoProvider.GetDacServices and remove the CustomDataTarget.DacSignatureVerificationEnabled
property.

Adds data target support for ICLRContractLocator for when clrmd loads the cdac.